It works!
First working version - sped up the I/O loop considerably and debugged issues with the hardware.  The startup delay is still too long to run off of the Atari's power so it has to be powered separately, and it requires a bus transeiver (eg: a 74LS245 hooked up to the inverted output of A12) to cut off the data lines when A12 is low.